refactor : Menu enum 클래스 따로 추출
